requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2018.8.24
chardet==3.0.4
colorama==0.4.4
coverage==4.5.1
coveralls==1.5.1
Django==2.2.10
django-cors-middleware==1.2.0
django-extensions==1.7.1
django-filter==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
factory-boy==2.11.1
Faker==0.9.2
idna==2.7
isort==5.6.4
itsdangerous==2.0.1
lazy-object-proxy==1.4.3
Markdown==3.0.1
mccabe==0.6.1
mysqlclient @ file:///E:/Study%20Material/Universal%20Informatics/Django/mysqlclient-1.4.6-cp38-cp38-win32.whl
pathlib==1.0.1
PyJWT==1.6.4
pylint==2.6.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.20.0
requests-cache==0.6.3
six==1.10.0
sqlparse==0.4.1
text-unidecode==1.2
toml==0.10.2
url-normalize==1.4.3
urllib3==1.24.2
wrapt==1.12.1